iot-cloud-project/requirements.txt
google-cloud-bigquery>=2.0.0
google-cloud-storage>=1.0.0
Flask>=2.0.0

